refactor: extract pure definition data and input-type helpers into leaf modules
Split the pure, side-effect-free parts of definitions.js into two stateless leaf modules so they can be imported (and unit-tested) in plain Node without pulling in the DOM, app state, or the WASM SDK.

input-types.js holds normalizeType and SUPPORTED_INPUT_TYPES, the input-type concept that form/collect.js and form/render.js both need. definitions-data.js holds the supported-operation sets, DPNS definitions, proof/auth/type config, getTypeConfig, and filterDefinitions, importing only input-types.js.

definitions.js is now just loadDefinitions (the one impure module that touches state, elements, EvoSDK, and fetch) plus re-exports of the pure symbols, so existing importers (main, execute, auth) are unchanged. form/collect.js and form/render.js are repointed at the leaves directly. normalizeType is removed from ui.js.

filterDefinitions, normalizeType, and the definition constants now import cleanly in Node; definitions.js still requires a DOM, as expected, since that is where the side effects live.
