+++ public/src/input-types.js
@@ -1,0 +1,23 @@
+export const SUPPORTED_INPUT_TYPES = new Set([
+  'text',
+  'string',
+  'textarea',
+  'number',
+  'checkbox',
+  'json',
+  'select',
+  'multiselect',
+  'array',
+  'button',
+  'dynamic',
+  'keyPreview',
+]);
+
+export function normalizeType(type) {
+  if (!type) return 'text';
+  switch (type) {
+    case 'string': return 'text';
+    case 'textarea': return 'textarea';
+    default: return type;
+  }
+}
